Use int instead of int64 in ThreadConfig sizes
Turns out if the _KiB/_MiB operators are consteval, MSVC warning about implicit cast doesn't trigger.

+++ Bedrock/Thread.h
@@ -24,10 +24,10 @@ enum class EThreadPriority : int8
 
 struct ThreadConfig
 {
-	String          mName        = "";                      // The thread name.
-	int64           mStackSize   = 128_KiB;					// The stack size of the thread.
-	int64           mTempMemSize = 1_MiB;					// Initialize a temp memory of that size of the thread. Can be 0.
-	EThreadPriority mPriority    = EThreadPriority::Normal;	// The priority of the thread.
+	String          mName        = "";      // The thread name.
+	int             mStackSize   = 128_KiB; // The stack size of the thread.
+	int             mTempMemSize = 1_MiB;   // Initialize a temp memory of that size of the thread. Can be 0.
+	EThreadPriority mPriority    = EThreadPriority::Normal; // The priority of the thread.
 };
